+++ src/engine/Collision/Colliders/CircleCollider.ts
@@ -45,14 +45,19 @@ export class CircleCollider extends Collider {
   }
 
   private _naturalRadius: number;
+
+  private _radius: number | undefined;
   /**
    * Get the radius of the circle
    */
   public get radius(): number {
+    if (this._radius) {
+      return this._radius;
+    }
     const tx = this._transform;
     const scale = tx?.globalScale ?? Vector.One;
     // This is a trade off, the alternative is retooling circles to support ellipse collisions
-    return this._naturalRadius * Math.min(scale.x, scale.y);
+    return (this._radius = this._naturalRadius * Math.min(scale.x, scale.y));
   }
 
   /**
@@ -63,6 +68,8 @@ export class CircleCollider extends Collider {
     const scale = tx?.globalScale ?? Vector.One;
     // This is a trade off, the alternative is retooling circles to support ellipse collisions
     this._naturalRadius = val / Math.min(scale.x, scale.y);
+    this._localBoundsDirty = true;
+    this._radius = val;
   }
 
   private _transform: Transform;
@@ -211,31 +218,20 @@ export class CircleCollider extends Collider {
    * Get the axis aligned bounding box for the circle collider in world coordinates
    */
   public get bounds(): BoundingBox {
-    const tx = this._transform;
-    const scale = tx?.globalScale ?? Vector.One;
-    const rotation = tx?.globalRotation ?? 0;
-    const pos = tx?.globalPos ?? Vector.Zero;
-    return new BoundingBox(
-      this.offset.x - this._naturalRadius,
-      this.offset.y - this._naturalRadius,
-      this.offset.x + this._naturalRadius,
-      this.offset.y + this._naturalRadius
-    )
-      .rotate(rotation)
-      .scale(scale)
-      .translate(pos);
+    return this.localBounds.transform(this._globalMatrix);
   }
 
+  private _localBoundsDirty = true;
+  private _localBounds: BoundingBox;
   /**
    * Get the axis aligned bounding box for the circle collider in local coordinates
    */
   public get localBounds(): BoundingBox {
-    return new BoundingBox(
-      this.offset.x - this._naturalRadius,
-      this.offset.y - this._naturalRadius,
-      this.offset.x + this._naturalRadius,
-      this.offset.y + this._naturalRadius
-    );
+    if (this._localBoundsDirty) {
+      this._localBounds = new BoundingBox(-this._naturalRadius, -this._naturalRadius, +this._naturalRadius, +this._naturalRadius);
+      this._localBoundsDirty = false;
+    }
+    return this._localBounds;
   }
 
   /**
@@ -259,6 +255,7 @@ export class CircleCollider extends Collider {
     const globalMat = transform.matrix ?? this._globalMatrix;
     globalMat.clone(this._globalMatrix);
     this._globalMatrix.translate(this.offset.x, this.offset.y);
+    this._radius = undefined;
   }
 
   /**