+++ src/engine/Collision/PhysicsWorld.ts
@@ -1,10 +1,19 @@
 import { Ray } from '../Math/ray';
 import { DeepRequired } from '../Util/Required';
 import { Observable } from '../Util/Observable';
-import { DynamicTreeCollisionProcessor, RayCastHit, RayCastOptions } from './Index';
+import {
+  BoundingBox,
+  Collider,
+  CollisionProcessor,
+  DynamicTreeCollisionProcessor,
+  RayCastHit,
+  RayCastOptions,
+  SparseHashGridCollisionProcessor
+} from './Index';
 import { BodyComponent } from './BodyComponent';
 import { PhysicsConfig } from './PhysicsConfig';
 import { watchDeep } from '../Util/Watch';
+import { Vector } from '../Math/vector';
 
 export class PhysicsWorld {
   $configUpdate = new Observable<DeepRequired<PhysicsConfig>>();
@@ -21,16 +30,20 @@ export class PhysicsWorld {
     this.$configUpdate.notifyAll(newConfig);
   }
 
-  private _collisionProcessor: DynamicTreeCollisionProcessor;
+  private _collisionProcessor: CollisionProcessor;
   /**
    * Spatial data structure for locating potential collision pairs and ray casts
    */
-  public get collisionProcessor(): DynamicTreeCollisionProcessor {
+  public get collisionProcessor(): CollisionProcessor {
     if (this._configDirty) {
       this._configDirty = false;
       // preserve tracked colliders if config updates
       const colliders = this._collisionProcessor.getColliders();
-      this._collisionProcessor = new DynamicTreeCollisionProcessor(this._config);
+      if (this._config.spatialPartition.type === 'sparse-hash-grid') {
+        this._collisionProcessor = new SparseHashGridCollisionProcessor(this._config.spatialPartition);
+      } else {
+        this._collisionProcessor = new DynamicTreeCollisionProcessor(this._config);
+      }
       for (const collider of colliders) {
         this._collisionProcessor.track(collider);
       }
@@ -43,7 +56,11 @@ export class PhysicsWorld {
       this._configDirty = true;
       BodyComponent.updateDefaultPhysicsConfig(config.bodies);
     });
-    this._collisionProcessor = new DynamicTreeCollisionProcessor(this.config);
+    if (this._config.spatialPartition.type === 'sparse-hash-grid') {
+      this._collisionProcessor = new SparseHashGridCollisionProcessor(this._config.spatialPartition);
+    } else {
+      this._collisionProcessor = new DynamicTreeCollisionProcessor(this._config);
+    }
   }
 
   /**
@@ -53,5 +70,16 @@ export class PhysicsWorld {
    */
   public rayCast(ray: Ray, options?: RayCastOptions): RayCastHit[] {
     return this.collisionProcessor.rayCast(ray, options);
+  }
+
+  /**
+   * Query for colliders in the scene's physics world
+   * @param point
+   */
+  public query(point: Vector): Collider[];
+  public query(bounds: BoundingBox): Collider[];
+  public query(pointOrBounds: Vector | BoundingBox): Collider[] {
+    // FIXME workaround TS: https://github.com/microsoft/TypeScript/issues/14107
+    return this._collisionProcessor.query(pointOrBounds as any);
   }
 }