+++ src/spec/CollisionSpec.ts
@@ -50,7 +50,16 @@ describe('A Collision', () => {
   });
 
   it('order of actors collision should not matter when an Active and Active Collision', () => {
-    const collisionTree = new ex.DynamicTreeCollisionProcessor(DefaultPhysicsConfig);
+    const collisionTree = new ex.DynamicTreeCollisionProcessor({
+      ...DefaultPhysicsConfig,
+      ...{
+        spatialPartition: {
+          type: 'dynamic-tree',
+          boundsPadding: 5,
+          velocityMultiplier: 2
+        }
+      }
+    });
 
     actor1.body.collisionType = ex.CollisionType.Active;
     actor2.body.collisionType = ex.CollisionType.Active;
@@ -67,7 +76,16 @@ describe('A Collision', () => {
   });
 
   it('order of actors collision should not matter when an Active and Passive Collision', () => {
-    const collisionTree = new ex.DynamicTreeCollisionProcessor(DefaultPhysicsConfig);
+    const collisionTree = new ex.DynamicTreeCollisionProcessor({
+      ...DefaultPhysicsConfig,
+      ...{
+        spatialPartition: {
+          type: 'dynamic-tree',
+          boundsPadding: 5,
+          velocityMultiplier: 2
+        }
+      }
+    });
 
     actor1.body.collisionType = ex.CollisionType.Active;
     actor2.body.collisionType = ex.CollisionType.Passive;
@@ -84,7 +102,7 @@ describe('A Collision', () => {
   });
 
   it('order of actors collision should not matter when an Active and PreventCollision', () => {
-    const collisionTree = new ex.DynamicTreeCollisionProcessor(DefaultPhysicsConfig);
+    const collisionTree = new ex.SparseHashGridCollisionProcessor({ size: 10 });
 
     actor1.body.collisionType = ex.CollisionType.Active;
     actor2.body.collisionType = ex.CollisionType.PreventCollision;
@@ -101,7 +119,7 @@ describe('A Collision', () => {
   });
 
   it('order of actors collision should not matter when an Active and Fixed', () => {
-    const collisionTree = new ex.DynamicTreeCollisionProcessor(DefaultPhysicsConfig);
+    const collisionTree = new ex.SparseHashGridCollisionProcessor({ size: 10 });
 
     actor1.body.collisionType = ex.CollisionType.Active;
     actor2.body.collisionType = ex.CollisionType.Fixed;
@@ -118,7 +136,7 @@ describe('A Collision', () => {
   });
 
   it('order of actors collision should not matter when an Fixed and Fixed', () => {
-    const collisionTree = new ex.DynamicTreeCollisionProcessor(DefaultPhysicsConfig);
+    const collisionTree = new ex.SparseHashGridCollisionProcessor({ size: 10 });
 
     actor1.body.collisionType = ex.CollisionType.Fixed;
     actor2.body.collisionType = ex.CollisionType.Fixed;